requirements.txt
ndspy>=4.2.0
pillow>=11.1.0
git+https://github.com/Xzonn/MachineTranslationTools.git
git+https://github.com/Xzonn/nitrogfx-py.git
